refac: 코드 분리, 폴더 이동
