+++ src/utils/core/render/createElementWithAttributes.ts
@@ -1,4 +1,5 @@
 import { IVDOMNode, TIntrinsicType, TVDOMProps } from "@/types/vdom";
+import { updateAttributes } from "@/utils/core/render/updateAttributes";
 
 const createElementWithAttributes = (normalizedNode: IVDOMNode<TVDOMProps>) => {
   // create element
@@ -9,30 +10,3 @@ const createElementWithAttributes = (normalizedNode: IVDOMNode<TVDOMProps>) => {
 };
 
 export default createElementWithAttributes;
-
-export const updateAttributes = (element: HTMLElement, props: TVDOMProps) => {
-  Object.entries(props).forEach(([key, value]) => {
-    if (key === "children") return;
-    if (key.startsWith("on") && typeof value === "function") {
-      let eventName = key.toLowerCase().slice(2);
-      if (eventName === "change") eventName = "input";
-      element.addEventListener(eventName, value);
-      return;
-    }
-    if (element instanceof HTMLInputElement) {
-      if (key === "checked") {
-        element.checked = value as boolean;
-        return;
-      }
-      if (key === "value") {
-        element.value = value as string;
-        return;
-      }
-    }
-    if (key === "style" && typeof value === "object") {
-      Object.assign(element.style, value);
-      return;
-    }
-    element.setAttribute(key, value as string);
-  });
-};
