+++ src/utils/core/render/rerender/commitDiffs.ts
@@ -1,16 +1,13 @@
-import { root } from "./rootManager";
-import { updateAttributes } from "./render/createElementWithAttributes";
-import { IDiffChange } from "./reconcile";
+import { root } from "@/utils/core/registry/root";
+import { updateAttributes } from "@/utils/core/render/updateAttributes";
+import { IDiffChange } from "@/utils/core/render/rerender/reconcile";
 import { IVDOMNode, TPrimitiveNode, TVDOMProps } from "@/types/vdom";
-import createElementWithAttributes from "./render/createElementWithAttributes";
-const isPrimitiveNode = (
-  node: IVDOMNode | TPrimitiveNode | null
-): node is TPrimitiveNode => {
-  return typeof node === "string" || typeof node === "number";
-};
+import createElementWithAttributes from "@/utils/core/render/createElementWithAttributes";
+import { isTextNode } from "@/utils/isTextNode";
+import { getChildrenToArray } from "@/utils/getChildrenToArray";
+
 const getElementByPath = (path: number[]): HTMLElement | null => {
   let current = root.getRootElement();
-
   for (const index of path) {
     if (!current) return null;
     const children = Array.from(current.childNodes);
@@ -21,38 +18,29 @@ const getElementByPath = (path: number[]): HTMLElement | null => {
   return current;
 };
 
-export const getParentElement = (path: number[]): HTMLElement | null => {
+const getParentElement = (path: number[]): HTMLElement | null => {
   if (path.length === 0) return root.getRootElement();
   return getElementByPath(path.slice(0, -1)) as HTMLElement;
 };
 
-export const createDOMNode = (
+const createDOMNode = (
   node: IVDOMNode<TVDOMProps> | TPrimitiveNode
 ): HTMLElement | Text => {
-  if (isPrimitiveNode(node)) {
-    return document.createTextNode(String(node));
-  }
+  if (isTextNode(node)) return document.createTextNode(String(node));
   return createElementWithAttributes(node);
 };
 
 const renderChildren = (node: IVDOMNode<TVDOMProps>, element: HTMLElement) => {
   if (!node.props.children) return;
-
-  const children = Array.isArray(node.props.children)
-    ? node.props.children
-    : [node.props.children];
-
+  const children = getChildrenToArray(node.props.children);
   children.forEach((child) => {
-    if (!child) return;
-
-    if (isPrimitiveNode(child)) {
+    if (isTextNode(child))
       element.appendChild(document.createTextNode(String(child)));
-    } else {
+    else {
       const childElement = createDOMNode(child);
       element.appendChild(childElement);
-      if (!isPrimitiveNode(child)) {
+      if (!isTextNode(child))
         renderChildren(child, childElement as HTMLElement);
-      }
     }
   });
 };
@@ -64,20 +52,14 @@ export const commitDiffs = (diffs: IDiffChange[]) => {
         if (!diff.newNode) return;
         const parent = getParentElement(diff.path);
         if (!parent) return;
-
-        const newElement = createDOMNode(diff.newNode);
-        if (!isPrimitiveNode(diff.newNode)) {
-          renderChildren(diff.newNode, newElement as HTMLElement);
-        }
-
+        const newElement = createDOMNode(diff.newNode) as HTMLElement;
+        if (!isTextNode(diff.newNode)) renderChildren(diff.newNode, newElement);
         const index = diff.path[diff.path.length - 1];
         const nextSibling = parent.childNodes[index];
 
-        if (nextSibling) {
-          parent.insertBefore(newElement, nextSibling);
-        } else {
-          parent.appendChild(newElement);
-        }
+        if (nextSibling) parent.insertBefore(newElement, nextSibling);
+        else parent.appendChild(newElement);
+
         break;
       }
       case "DELETE": {
@@ -95,15 +77,12 @@ export const commitDiffs = (diffs: IDiffChange[]) => {
         break;
       }
       case "UPDATE": {
-        if (!diff.newNode || isPrimitiveNode(diff.newNode)) return;
+        if (!diff.newNode || isTextNode(diff.newNode)) return;
         const targetElement = getElementByPath(diff.path);
         if (!targetElement) return;
-
         // 이벤트 리스너 정리
         const oldProps =
-          diff.oldNode && !isPrimitiveNode(diff.oldNode)
-            ? diff.oldNode.props
-            : {};
+          diff.oldNode && !isTextNode(diff.oldNode) ? diff.oldNode.props : {};
         Object.entries(oldProps).forEach(([key, value]) => {
           if (key.startsWith("on") && typeof value === "function") {
             const eventName = key.toLowerCase().slice(2);
