+++ src/utils/getChildrenToArray.ts
@@ -1,0 +1,3 @@
+export const getChildrenToArray = <T>(children: T | T[]): T[] => {
+  return Array.isArray(children) ? children : [children];
+};
